refactor(valuer)：separate row and collection
