+++ internal/topo/operator/aggregate_operator.go
@@ -1,4 +1,4 @@
-// Copyright 2021 EMQ Technologies Co., Ltd.
+// Copyright 2021-2022 EMQ Technologies Co., Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -35,17 +35,17 @@ func (p *AggregateOp) Apply(ctx api.StreamContext, data interface{}, fv *xsql.Fu
 	grouped := data
 	var wr *xsql.WindowRange
 	if p.Dimensions != nil {
-		var ms []xsql.DataValuer
+		var ms []xsql.Row
 		switch input := data.(type) {
 		case error:
 			return input
-		case xsql.DataValuer:
+		case xsql.Row:
 			ms = append(ms, input)
 		case xsql.WindowTuplesSet:
 			if len(input.Content) != 1 {
 				return fmt.Errorf("run Group By error: the input WindowTuplesSet with multiple tuples cannot be evaluated")
 			}
-			ms = make([]xsql.DataValuer, len(input.Content[0].Tuples))
+			ms = make([]xsql.Row, len(input.Content[0].Tuples))
 			for i, m := range input.Content[0].Tuples {
 				//this is needed or it will always point to the last
 				t := m
@@ -53,7 +53,7 @@ func (p *AggregateOp) Apply(ctx api.StreamContext, data interface{}, fv *xsql.Fu
 			}
 			wr = input.WindowRange
 		case *xsql.JoinTupleSets:
-			ms = make([]xsql.DataValuer, len(input.Content))
+			ms = make([]xsql.Row, len(input.Content))
 			for i, m := range input.Content {
 				t := m
 				ms[i] = &t
@@ -76,7 +76,7 @@ func (p *AggregateOp) Apply(ctx api.StreamContext, data interface{}, fv *xsql.Fu
 				}
 			}
 			if ts, ok := result[name]; !ok {
-				result[name] = &xsql.GroupedTuples{Content: []xsql.DataValuer{m}, WindowRange: wr}
+				result[name] = &xsql.GroupedTuples{Content: []xsql.Row{m}, WindowRange: wr}
 			} else {
 				ts.Content = append(ts.Content, m)
 			}
